Fix bug when syncing an unmanaged directory symlink
Closes #2

+++ dotfiles/core.py
@@ -38,7 +38,7 @@ class Dotfile(object):
             if not force:
                 print "Skipping \"%s\", use --force to override" % self.basename
                 return
-            if os.path.isdir(self.name):
+            if os.path.isdir(self.name) and not os.path.islink(self.name):
                 shutil.rmtree(self.name)
             else:
                 os.remove(self.name)
